add header util
